--- src/main/java/me/ryder/cmds/CmdBountyList.java
@@ -1,4 +1,0 @@
-package me.ryder.cmds;
-
-public class CmdBountyList {
-}
