+++ src/main/java/me/ryder/headhunters/cmds/CmdBountyList.java
@@ -1,0 +1,4 @@
+package me.ryder.headhunters.cmds;
+
+public class CmdBountyList {
+}
