+++ src/main/java/me/ryder/headhunters/cmds/CmdBountyReload.java
@@ -1,4 +1,4 @@
-package me.ryder.cmds;
+package me.ryder.headhunters.cmds;
 
 public class CmdBountyReload {
 }
