+++ src/main/java/me/ryder/headhunters/util/fm/Cortex.java
@@ -1,0 +1,4 @@
+package me.ryder.headhunters.util.fm;
+
+public class Cortex {
+}
