+++ src/main/java/me/ryder/headhunters/util/fm/FileManager.java
@@ -1,4 +1,4 @@
-package me.ryder.util.fm;
+package me.ryder.headhunters.util.fm;
 
 import java.io.File;
 import java.io.FileOutputStream;
