--- src/main/java/me/ryder/util/fm/Cortex.java
@@ -1,4 +1,0 @@
-package me.ryder.util.fm;
-
-public class Cortex {
-}
